feat: refactor Jarvis backend structure and implement asynchronous command registry system
