+++ backend/jarvis/commands/registry.py
