+++ backend/jarvis/commands/vision.py
