+++ docker-compose.yml
@@ -2,14 +2,14 @@ services:
 
   backend:
     build:
-      context: .
-      dockerfile: Dockerfile.backend
+      context: ./backend
+      dockerfile: Dockerfile
     ports:
       - "8000:8000"
     env_file:
       - .env
     volumes:
-      - .:/app
+      - ./backend:/app
     restart: always
 
   frontend:
